requirements.txt
requests
beautifulsoup4
selenium
pandas
google-generativeai
lxml